+++ Packages/BrandonUtils/Standalone/Enums/Clusivity.cs
@@ -1,0 +1,6 @@
+﻿namespace BrandonUtils.Standalone.Enums {
+    public enum Clusivity {
+        Inclusive,
+        Exclusive,
+    }
+}