+++ Packages/BrandonUtils/Tests/Standalone/BloopTests.cs
@@ -8,6 +8,8 @@ using BrandonUtils.Testing;
 
 using NUnit.Framework;
 
+using Is = NUnit.Framework.Is;
+
 namespace BrandonUtils.Tests.Standalone {
     /// <summary>
     /// Tests for <see cref="Bloop"/>s