+++ Packages/BrandonUtils/Tests/Standalone/BrandomTests.cs
@@ -9,6 +9,8 @@ using BrandonUtils.Testing;
 
 using NUnit.Framework;
 
+using Is = BrandonUtils.Testing.Is;
+
 namespace BrandonUtils.Tests.Standalone {
     public class BrandomTests {
         [Test]