+++ Packages/BrandonUtils/Tests/Standalone/Collections/FailableTests.cs
@@ -6,6 +6,8 @@ using BrandonUtils.Testing;
 
 using NUnit.Framework;
 
+using Is = NUnit.Framework.Is;
+
 namespace BrandonUtils.Tests.Standalone.Collections {
     public class FailableTests {
         #region Example methods