+++ Packages/BrandonUtils/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.1] - 7/5/2021
+
+### Added
+
+- `Clusivity` enum
+- `ClusivityUtils`
+- `Is`, which is an "extension" of NUnit's `Is` but includes my custom assertions
+
+### Changed
+
+- Improved `ApproximationConstraint`
+    - Moved default value setting into `ApproximationConstraint` constructors
+    - Removed generic factory methods
+    - Used new fancy new `Coercively` class
+    - Added ability to specify the `Clusivity` of the range
+
 ## [2.6.0] - 7/4/2021
 
 ### Added